+++ tests/test_decisiontree.py
@@ -1,5 +1,5 @@
 import unittest
-from decisiontree.decisiontree import DecisionTree
+from algorithms.decisiontree import DecisionTree
 import pandas as pd
 from sklearn.datasets import load_iris
